% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_stop_data_collection_by_agent_ids}
\alias{applicationdiscoveryservice_stop_data_collection_by_agent_ids}
\title{Instructs the specified agents or connectors to stop collecting data}
\usage{
applicationdiscoveryservice_stop_data_collection_by_agent_ids(agentIds)
}
\arguments{
\item{agentIds}{[required] The IDs of the agents or connectors from which to stop collecting data.}
}
\description{
Instructs the specified agents or connectors to stop collecting data.
}
\section{Request syntax}{
\preformatted{svc$stop_data_collection_by_agent_ids(
  agentIds = list(
    "string"
  )
)
}
}

\keyword{internal}
