% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_endpoint}
\alias{databasemigrationservice_delete_endpoint}
\title{Deletes the specified endpoint}
\usage{
databasemigrationservice_delete_endpoint(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
endpoint.}
}
\description{
Deletes the specified endpoint.
}
\details{
All tasks associated with the endpoint must be deleted before you can
delete the endpoint.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$delete_endpoint(
  EndpointArn = "string"
)
}
}

\examples{
# Deletes the specified endpoint. All tasks associated with the endpoint
# must be deleted before you can delete the endpoint.
# 
\donttest{databasemigrationservice$delete_endpoint(
  EndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:RAAR3R22XSH46S3PWLC3NJAWKM"
)}

}
\keyword{internal}
