% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_event_subscriptions}
\alias{databasemigrationservice_describe_event_subscriptions}
\title{Lists all the event subscriptions for a customer account}
\usage{
databasemigrationservice_describe_event_subscriptions(SubscriptionName,
  Filters, MaxRecords, Marker)
}
\arguments{
\item{SubscriptionName}{The name of the AWS DMS event subscription to be described.}

\item{Filters}{Filters applied to the action.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Lists all the event subscriptions for a customer account. The
description of a subscription includes \code{SubscriptionName},
\code{SNSTopicARN}, \code{CustomerID}, \code{SourceType}, \code{SourceID}, \code{CreationTime},
and \code{Status}.
}
\details{
If you specify \code{SubscriptionName}, this action lists the description for
that subscription.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$describe_event_subscriptions(
  SubscriptionName = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

\keyword{internal}
