% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_agent}
\alias{datasync_create_agent}
\title{Activates an AWS DataSync agent that you have deployed on your host}
\usage{
datasync_create_agent(ActivationKey, AgentName, Tags)
}
\arguments{
\item{ActivationKey}{[required] Your agent activation key. You can get the activation key either by
sending an HTTP GET request with redirects that enable you to get the
agent IP address (port 80). Alternatively, you can get it from the AWS
DataSync console.

The redirect URL returned in the response provides you the activation
key for your agent in the query string parameter \code{activationKey}. It
might also include other activation-related parameters; however, these
are merely defaults. The arguments you pass to this API call determine
the actual configuration of your agent. For more information, see
Activating a Sync Agent in the \emph{AWS DataSync User Guide.}}

\item{AgentName}{The name you configured for your agent. This value is a text reference
that is used to identify the agent in the console.}

\item{Tags}{The key-value pair that represents the tag you want to associate with
the agent. The value can be an empty string. This value helps you
manage, filter, and search for your agents.

Valid characters for key and value are letters, spaces, and numbers
representable in UTF-8 format, and the following special characters: + -
= . \_ : / @.}
}
\description{
Activates an AWS DataSync agent that you have deployed on your host. The
activation process associates your agent with your account. In the
activation process, you specify information such as the AWS Region that
you want to activate the agent in. You activate the agent in the AWS
Region where your target locations (in Amazon S3 or Amazon EFS) reside.
Your tasks are created in this AWS Region.
}
\details{
You can use an agent for more than one location. If a task uses multiple
agents, all of them need to have status AVAILABLE for the task to run.
If you use multiple agents for a source location, the status of all the
agents must be AVAILABLE for the task to run. For more information, see
Activating a Sync Agent in the \emph{AWS DataSync User Guide.}

Agents are automatically updated by AWS on a regular basis, using a
mechanism that ensures minimal interruption to your tasks.
}
\section{Request syntax}{
\preformatted{svc$create_agent(
  ActivationKey = "string",
  AgentName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
