% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_task}
\alias{datasync_update_task}
\title{Updates the metadata associated with a task}
\usage{
datasync_update_task(TaskArn, Options, Name)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the resource name of the task to
update.}

\item{Options}{}

\item{Name}{The name of the task to update.}
}
\description{
Updates the metadata associated with a task.
}
\section{Request syntax}{
\preformatted{svc$update_task(
  TaskArn = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"NONE",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"BEST_EFFORT"|"PRESERVE",
    BytesPerSecond = 123
  ),
  Name = "string"
)
}
}

\keyword{internal}
