% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{get_packages}
\alias{get_packages}
\title{Generate txt file with all installed packages}
\usage{
get_packages(file_name)
}
\arguments{
\item{file_name}{Character, desired file name as txt}
}
\value{
A txt file named "package_list.txt" saved to your working directory
}
\description{
Generate txt file with all installed packages
}
