% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_active.R
\name{git_active}
\alias{git_active}
\title{Determine whether git is active in active project}
\usage{
git_active()
}
\value{
Logical value, whether git is installed in active project
}
\description{
Determine whether git is active in active project
}
