% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_delete_app_launch_configuration}
\alias{sms_delete_app_launch_configuration}
\title{Deletes existing launch configuration for an application}
\usage{
sms_delete_app_launch_configuration(appId)
}
\arguments{
\item{appId}{ID of the application associated with the launch configuration.}
}
\description{
Deletes existing launch configuration for an application.
}
\section{Request syntax}{
\preformatted{sms$delete_app_launch_configuration(
  appId = "string"
)
}
}

\keyword{internal}
