% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_launch_configuration}
\alias{sms_get_app_launch_configuration}
\title{Retrieves the application launch configuration associated with an
application}
\usage{
sms_get_app_launch_configuration(appId)
}
\arguments{
\item{appId}{ID of the application launch configuration.}
}
\description{
Retrieves the application launch configuration associated with an
application.
}
\section{Request syntax}{
\preformatted{sms$get_app_launch_configuration(
  appId = "string"
)
}
}

\keyword{internal}
