% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_create_job}
\alias{snowball_create_job}
\title{Creates a job to import or export data between Amazon S3 and your
on-premises data center}
\usage{
snowball_create_job(JobType, Resources, Description, AddressId,
  KmsKeyARN, RoleARN, SnowballCapacityPreference, ShippingOption,
  Notification, ClusterId, SnowballType, ForwardingAddressId)
}
\arguments{
\item{JobType}{Defines the type of job that you're creating.}

\item{Resources}{Defines the Amazon S3 buckets associated with this job.

With \code{IMPORT} jobs, you specify the bucket or buckets that your
transferred data will be imported into.

With \code{EXPORT} jobs, you specify the bucket or buckets that your
transferred data will be exported from. Optionally, you can also specify
a \code{KeyRange} value. If you choose to export a range, you define the
length of the range by providing either an inclusive \code{BeginMarker}
value, an inclusive \code{EndMarker} value, or both. Ranges are UTF-8 binary
sorted.}

\item{Description}{Defines an optional description of this specific job, for example
\code{Important Photos 2016-08-11}.}

\item{AddressId}{The ID for the address that you want the Snowball shipped to.}

\item{KmsKeyARN}{The \code{KmsKeyARN} that you want to associate with this job. \code{KmsKeyARN}s
are created using the
\href{http://docs.aws.amazon.com/kms/latest/APIReference/API_CreateKey.html}{CreateKey}
AWS Key Management Service (KMS) API action.}

\item{RoleARN}{The \code{RoleARN} that you want to associate with this job. \code{RoleArn}s are
created using the
\href{http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html}{CreateRole}
AWS Identity and Access Management (IAM) API action.}

\item{SnowballCapacityPreference}{If your job is being created in one of the US regions, you have the
option of specifying what size Snowball you'd like for this job. In all
other regions, Snowballs come with 80 TB in storage capacity.}

\item{ShippingOption}{The shipping speed for this job. This speed doesn't dictate how soon
you'll get the Snowball, rather it represents how quickly the Snowball
moves to its destination while in transit. Regional shipping speeds are
as follows:
\itemize{
\item In Australia, you have access to express shipping. Typically,
Snowballs shipped express are delivered in about a day.
\item In the European Union (EU), you have access to express shipping.
Typically, Snowballs shipped express are delivered in about a day.
In addition, most countries in the EU have access to standard
shipping, which typically takes less than a week, one way.
\item In India, Snowballs are delivered in one to seven days.
\item In the US, you have access to one-day shipping and two-day shipping.
}}

\item{Notification}{Defines the Amazon Simple Notification Service (Amazon SNS) notification
settings for this job.}

\item{ClusterId}{The ID of a cluster. If you're creating a job for a node in a cluster,
you need to provide only this \code{clusterId} value. The other job
attributes are inherited from the cluster.}

\item{SnowballType}{The type of AWS Snowball device to use for this job. The only supported
device types for cluster jobs are \code{EDGE}, \code{EDGE_C}, and \code{EDGE_CG}.}

\item{ForwardingAddressId}{The forwarding address ID for a job. This field is not supported in most
regions.}
}
\description{
Creates a job to import or export data between Amazon S3 and your
on-premises data center. Your AWS account must have the right trust
policies and permissions in place to create a job for Snowball. If
you're creating a job for a node in a cluster, you only need to provide
the \code{clusterId} value; the other job attributes are inherited from the
cluster.
}
\section{Request syntax}{
\preformatted{svc$create_job(
  JobType = "IMPORT"|"EXPORT"|"LOCAL_USE",
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "string",
        KeyRange = list(
          BeginMarker = "string",
          EndMarker = "string"
        )
      )
    ),
    LambdaResources = list(
      list(
        LambdaArn = "string",
        EventTriggers = list(
          list(
            EventResourceARN = "string"
          )
        )
      )
    ),
    Ec2AmiResources = list(
      list(
        AmiId = "string",
        SnowballAmiId = "string"
      )
    )
  ),
  Description = "string",
  AddressId = "string",
  KmsKeyARN = "string",
  RoleARN = "string",
  SnowballCapacityPreference = "T50"|"T80"|"T100"|"T42"|"NoPreference",
  ShippingOption = "SECOND_DAY"|"NEXT_DAY"|"EXPRESS"|"STANDARD",
  Notification = list(
    SnsTopicARN = "string",
    JobStatesToNotify = list(
      "New"|"PreparingAppliance"|"PreparingShipment"|"InTransitToCustomer"|"WithCustomer"|"InTransitToAWS"|"WithAWSSortingFacility"|"WithAWS"|"InProgress"|"Complete"|"Cancelled"|"Listing"|"Pending"
    ),
    NotifyAll = TRUE|FALSE
  ),
  ClusterId = "string",
  SnowballType = "STANDARD"|"EDGE"|"EDGE_C"|"EDGE_CG",
  ForwardingAddressId = "string"
)
}
}

\examples{
# Creates a job to import or export data between Amazon S3 and your
# on-premises data center. Your AWS account must have the right trust
# policies and permissions in place to create a job for Snowball. If
# you're creating a job for a node in a cluster, you only need to provide
# the clusterId value; the other job attributes are inherited from the
# cluster.
\donttest{svc$create_job(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b",
  Description = "My Job",
  JobType = "IMPORT",
  KmsKeyARN = "arn:aws:kms:us-east-1:123456789012:key/abcd1234-12ab-34cd-56ef-123456123456",
  Notification = list(
    JobStatesToNotify = list(),
    NotifyAll = FALSE
  ),
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "arn:aws:s3:::MyBucket",
        KeyRange = structure(
          list(),
          .Names = character(
            0
          )
        )
      )
    )
  ),
  RoleARN = "arn:aws:iam::123456789012:role/snowball-import-S3-role",
  ShippingOption = "SECOND_DAY",
  SnowballCapacityPreference = "T80",
  SnowballType = "STANDARD"
)}

}
\keyword{internal}
